EY-1844: Pakkenamn
