+++ apps/etterlatte-migrering/src/main/kotlin/migrering/Pesyssak.kt
@@ -1,0 +1,15 @@
+package no.nav.etterlatte.migrering
+
+import no.nav.etterlatte.libs.common.behandling.Persongalleri
+import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
+import rapidsandrivers.migrering.Enhet
+import rapidsandrivers.migrering.PesysId
+import java.time.LocalDateTime
+
+data class Pesyssak(
+    val id: PesysId,
+    val enhet: Enhet,
+    val folkeregisteridentifikator: Folkeregisteridentifikator,
+    val mottattdato: LocalDateTime,
+    val persongalleri: Persongalleri
+)
